pyproject.toml
[project]
name = "myobservatory-testing-project"
version = "0.1.0"
description = "Comprehensive testing framework for MyObservatory mobile app and API using Appium, pytest-bdd, and requests"
authors = [
    {name = "Hailei Song", email = "hailei.song@foxmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "appium-python-client>=3.1.0",
    "pytest>=8.0.0",
    "pytest-bdd>=7.0.0",
    "pytest-html>=4.1.0",
    "pytest-xdist>=3.5.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "selenium>=4.16.0",
    "pillow>=10.2.0",
    "allure-pytest>=2.13.2",
    "tenacity>=8.2.3",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "config", "drivers", "pages", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--capture=no",
]
markers = [
    "android: Tests for Android platform",
    "ios: Tests for iOS platform",
    "api: API tests",
    "mobile: Mobile app tests",
    "smoke: Smoke tests",
    "regression: Regression tests",
    "slow: Slow running tests",
]
bdd_features_base_dir = "features"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
